refactor: move FIM constants to a dedicated file for better organization
